Add auton package and move commands inside
